hockeyclubs-pipeline/requirements.txt
Pillow>=10.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
openpyxl>=3.1.0
requests>=2.31.0
supabase>=2.0.0
python-dotenv>=1.0.0
